Do not convert double to int before displaying in _3D.

+++ src/main/scala/acumen/ui/threeD/ThreeDData.scala
@@ -116,7 +116,7 @@ class ThreeDData extends Publisher {
     value match {
       case VLit(GStr(s)) => _3DText = s
       case VLit(GInt(i)) => _3DText = i.toString
-      case VLit(GDouble(i)) => _3DText = i.toInt.toString
+      case VLit(GDouble(i)) => _3DText = i.toString
       case _ => throw _3DNameError(value)
     }
   }
